Completado leccion 314

+++ src/components/Product.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 import { useDispatch } from 'react-redux';
 import { productDeleteAction } from '../actions/productAction';
@@ -12,9 +13,20 @@ const Product = ({product}) => {
     const deleteProductConfirm = id => {
 
         // Preguntar al usuario
-
-        // Enviarlo al Action
-        dispatch( productDeleteAction(id) );
+        Swal.fire({
+            title: 'Atención',
+            text: "¿Desea eliminar el producto?",
+            icon: 'warning',
+            showCancelButton: true,
+            cancelButtonColor: '#3085d6',
+            confirmButtonColor: '#d33',
+            confirmButtonText: 'Si',
+            cancelButtonText: 'No, mantener el Producto'
+        }).then((result) => {
+            if(result.value){
+              dispatch( productDeleteAction(id) );  
+            }            
+        });        
     }
 
     return (
